chore: address issue #151
